+++ R/dump_download.R
@@ -2,8 +2,9 @@
 #'
 #' @param res_id a resource ID
 #' @return dataframe containing resource records
-#'
-dump_download <- function(res_id) {
+#' @keywords internal
+#' @noRd
+dump_download <- function(res_id, call = rlang::caller_env()) {
   # fetch the data
   content <- suppressMessages(
     phs_GET("dump", res_id)
@@ -11,9 +12,12 @@ dump_download <- function(res_id) {
 
   # if content is a web page
   if ("xml_document" %in% class(content)) {
-    cli::cli_abort(c(
-      "Can't find resource with ID {.var {res_id}} in datastore."
-    ))
+    cli::cli_abort(
+      c(
+        "Can't find resource with ID {.var {res_id}} in datastore."
+      ),
+      call = call
+    )
   }
 
   # return data