+++ R/suggest_dataset_name.R
@@ -1,7 +1,9 @@
 #' Throw an error with suggested dataset name, if possible.
 #'
 #' @inheritParams get_dataset
-suggest_dataset_name <- function(dataset_name) {
+#' @keywords internal
+#' @noRd
+suggest_dataset_name <- function(dataset_name, call = rlang::caller_env()) {
   content <- phs_GET("package_list", "")
 
   dataset_names <- unlist(content$result)
@@ -11,20 +13,28 @@ suggest_dataset_name <- function(dataset_name) {
 
   # if min distance is too big, abort
   if (min(string_distances) > 10) {
-    cli::cli_abort(c(
-      "Can't find the dataset name
+    cli::cli_abort(
+      c(
+        "Can't find the dataset name
       {.var {dataset_name}}, or a close match.",
-      i = "Find a dataset's name in the URL
+        i = "Find a dataset's name in the URL
       of its page on {.url www.opendata.nhs.scot.}"
-    ))
+      ),
+      call = call
+    )
   }
 
   # find closet match
-  closest_match <- dataset_names[which(string_distances == min(string_distances))]
+  closest_match <- dataset_names[which(
+    string_distances == min(string_distances)
+  )]
 
   # throw error with suggestion
-  cli::cli_abort(c(
-    "Can't find the dataset name {.var {dataset_name}}.",
-    "i" = "Did you mean {?any of }{.val {closest_match}}?"
-  ))
+  cli::cli_abort(
+    c(
+      "Can't find the dataset name {.var {dataset_name}}.",
+      "i" = "Did you mean {?any of }{.val {closest_match}}?"
+    ),
+    call = call
+  )
 }